1)TicketsUserController: 	create punch 	generateTempCode
2)override tempCode and createdTempCode after create punch
3)remove nullable from id

+++ BLL/Versions/V1/BusinessLogic/StoreBL.cs
@@ -64,7 +64,7 @@ namespace BLL.Versions.V1.BusinessLogic
                 return new BadRequestResult();
             }
 
-            ActionResult<Store> action = await storeDA.GetStore((long)storeDTO.Id);
+            ActionResult<Store> action = await storeDA.GetStore(storeDTO.Id);
             if(action == null || action.Value == null)
             {
                 return new NotFoundResult();
